requirements.txt
anyio==3.6.2
asgiref==3.6.0
async-timeout==4.0.2
autopep8==2.0.2
black==23.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.5
django-cors-headers==3.13.0
django-prometheus==2.2.0
django-rabbitmq==0.1.3
django-redis==5.2.0
django-rest-swagger==2.2.0
django-silk==5.0.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.4
gprof2dot==2022.7.29
gunicorn==21.2.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.1
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
openapi-codec==1.3.2
packaging==23.0
pika==1.3.1
prometheus-client==0.16.0
psycopg2-binary==2.9.1
python-dotenv==1.0.0
pycodestyle==2.10.0
PyJWT==2.6.0
pytz==2022.7.1
PyYAML==6.0
python-dateutil==2.8.2
redis==4.5.1
requests==2.28.2
rfc3986==1.5.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
sentry-sdk==1.3.1
simplejson==3.18.3
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.3
structlog==22.3.0
swagger-ui-py==22.7.13
tomli==2.0.1
uritemplate==4.1.1
urllib3==1.26.14
grpcio==1.56.2	
grpcio-tools==1.56.2
protobuf==4.23.4